Show 2 decimal digits in totals
[no ci]

+++ src/Common/TotalData.cs
@@ -1,4 +1,4 @@
 ﻿namespace myenergy.Common
 {
-    public record TotalData(int TotalProduction, int TotalNetInjection, int TotalNetUsage, int TotalUsage);
+    public record TotalData(double TotalProduction, double TotalNetInjection, double TotalNetUsage, double TotalUsage);
 }
